create file category and files

+++ database/migrations/2018_04_09_163005_create_file_tags_table.php
@@ -17,7 +17,7 @@ class CreateFileTagsTable extends Migration
             $table->increments('id');
             $table->integer('file_id')->unsigned();
             $table->foreign('file_id')->references('id')->on('files_info')->onDelete('cascade')->onUpdate('cascade');
-            $table->integer('tag_id')->unsigned();
+            $table->bigIncrements('tag_id')->unsigned();
             $table->foreign('tag_id')->references('id')->on('tags')->onDelete('cascade')->onUpdate('cascade');
         });
     }
